requirements.txt
alembic==0.9.6
attrs==17.4.0
bcrypt==3.1.4
bleach==1.5.0
-e git+https://github.com/pybox2d/pybox2d.git@055428c5ad3fab7faa80306883b806cc536c1ae9#egg=Box2D
certifi==2017.11.5
cffi==1.11.2
chardet==3.0.4
click==6.7
coverage==4.4.2
cycler==0.10.0
decorator==4.1.2
docker==2.7.0
docker-pycreds==0.2.1
enum34==1.1.6
Flask==0.12.2
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.3
Flask-Migrate==2.1.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
Flask-Testing==0.7.1
future==0.16.0
GitHub-Flask==3.2.0
gunicorn==19.7.1
-e git+https://github.com/openai/gym.git@0eccce8146d011f7691f4aff5766be7c5a4d8cbd#egg=gym
html5lib==0.9999999
idna==2.6
itsdangerous==0.24
Jinja2==2.10
Mako==1.0.7
Markdown==2.6.11
MarkupSafe==1.0
matplotlib==2.1.1
mysql-connector==2.1.4
networkx==2.0
numpy==1.14.0
Pillow==5.0.0
pluggy==0.6.0
protobuf==3.5.1
psycopg2==2.7.3.2
py==1.5.2
py-bcrypt==0.4
pycparser==2.18
pyglet==1.3.0
pyparsing==2.2.0
pytest==3.3.2
pytest-cov==2.5.1
pytest-flask==0.10.0
python-dateutil==2.6.1
python-editor==1.0.3
pytz==2017.3
PyWavelets==0.5.2
requests==2.18.4
ruamel.yaml==0.15.35
scikit-image==0.13.1
scipy==0.18.1
six==1.11.0
SQLAlchemy==1.2.0
tensorflow==1.4.1
tensorflow-tensorboard==0.4.0rc3
urllib3==1.22
webencodings==0.5.1
websocket-client==0.46.0
Werkzeug==0.14.1
